requirements.txt
pillow
django-admin-interface
django-crispy-forms
cx_oracle
requests
django-cors-headers
python-decouple
django-admin